code refactor: closes https://github.com/metanorma/mn-templates-iso/issues/12
